Add support for different height input grids
It will probably be useful to support different grid heights i.e.
resolutions at different network depths, similar to convnets and
hierarchical vision transformers. Update `ColumnAttention` and
`WiringCost` to allow this.

Also add an offset in the z direction to the wiring distance weight so
that no connection is free.