Add factorized linear classification head
Because we expect/hope for some non-uniformity in the final
representation map, standard global average pooling doesn't make as much
sense.

Instead, we're trying out a factorized linear classification head where
the weights for each class are factorized into an independent spatial
pooling map and feature projection, which should allow different classes
to attend to different regions of the map.